# openapi.yaml
openapi: 3.0.3
info:
  title: Fundor API (Target Specification)
  version: 2.3.0
  description: |
    **Target specification for the Fundor API**, providing grant/funding intelligence for Hungarian SMEs.
    This is the contract to implement in the Laravel application, not documentation of currently available endpoints.
    This project must implement the operations and schemas below. `x-frontend` records intended screen consumers;
    it does not imply that a screen or operation is already implemented.

    ## Laravel implementation (2026-09-21)

    The operations below are registered in `routes/api.php`, backed by database persistence and API-specific
    `hunter_session` cookies. Apply migrations before using them. Existing Blade login sessions are separate.
    See `docs/FRONTEND-API.md` for local setup, request examples and deployment configuration.

    `GET /api/health` remains an operational endpoint outside this contract. `GET /api/opportunities` is a
    compatibility alias of `GET /catalog` and enforces the same paywall.
    Refresh requires `FUNDOR_CATALOG_FEED_URL`, a deployment-controlled normalized catalog feed; without it,
    `POST /refresh` returns `503 REFRESH_UNAVAILABLE`. The raw EC portal ingestion adapter is not included.
    Subscription prices are deployment configuration because this contract does not specify numeric prices.
    Until configured, plan prices are absent and price-dependent CRM values are null, not fabricated revenue.
...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - {}
  - cookieAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: Create an account and sign in
      description: |
        Creates the account, opens a session (sets the cookie) and, if a lead with the same `email` exists, marks that lead
        as converted. `password` must be ≥ 4 chars (see implementation requirements); `username` matches `^[a-zA-Z0-9._-]{3,32}$`.
      x-frontend: { used-by: [Register form] }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, pattern: '^[a-zA-Z0-9._-]{3,32}$', example: acme }
                password: { type: string, format: password, minLength: 4 }
                company: { type: string }
                email: { type: string, format: email }
      responses:
        '201':
          description: Created and signed in.
          headers:
            Set-Cookie: { $ref: '#/components/headers/SetSessionCookie' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthActionResponse' }
        '400':
          description: "`INVALID_USERNAME` or `WEAK_PASSWORD`."
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '409':
          description: "`USERNAME_TAKEN`."
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Sign in
      x-frontend: { used-by: [Login form] }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        '200':
          description: Signed in.
          headers:
            Set-Cookie: { $ref: '#/components/headers/SetSessionCookie' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthActionResponse' }
        '401':
          description: "`BAD_CREDENTIALS`. The same answer for an unknown user and a wrong password."
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '403':
          description: "`ACCOUNT_DISABLED`."
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: hunter_session
      description: httpOnly, SameSite=Lax, 7-day sliding expiry. Set by `/auth/login` and `/auth/register`.
  headers:
    SetSessionCookie:
      description: "`hunter_session=<token>; Path=/; HttpOnly; SameSite=Lax; Max-Age=604800` (+ `Secure` when configured)."
      schema: { type: string }
...
  schemas:
...
    Error:
      type: object
      required: [error, code]
      properties:
        error: { type: string, description: 'A Hungarian message — a fallback only.' }
        code: { $ref: '#/components/schemas/ErrorCode' }
    ErrorCode:
      type: string
      description: Stable identifiers the frontend translates. Add new ones to the frontend's `errors.json` when introduced.
      enum:
        - INVALID_REQUEST
        - INTERNAL_ERROR
        - JSON_REQUIRED
        - CSRF_REJECTED
        - EMAIL_TAKEN
        - REFRESH_UNAVAILABLE
        - LOGIN_REQUIRED
        - ADMIN_REQUIRED
        - BAD_CREDENTIALS
        - ACCOUNT_DISABLED
        - USERNAME_TAKEN
        - INVALID_USERNAME
        - WEAK_PASSWORD
        - WRONG_PASSWORD
        - NO_SUCH_OPPORTUNITY
        - FIELD_REQUIRED
        - NO_SUCH_VERSION
        - NO_SUCH_USER
        - UNKNOWN_PLAN
        - CANNOT_DEMOTE_SELF
        - CANNOT_DELETE_SELF
        - EMAIL_REQUIRED
        - INVALID_EMAIL
        - CONSENT_REQUIRED
        - NO_SUCH_CONTACT
        - NO_SUCH_LEAD
        - NO_SUCH_NOTE
        - NO_SUCH_TASK
        - UNKNOWN_STAGE
        - EMPTY_NOTE
        - EMPTY_TASK
        - INVALID_DUE_DATE
...
    Entitlements:
      type: object
      required: [tier, explanations, calculator, applyLinks, exportData, admin]
      properties:
        tier: { type: string, enum: [anonymous, registered, subscriber, admin] }
        maxResults:
          type: number
          nullable: true
          description: How many *full* results a caller may see. `0` for anonymous/registered; **`null` = unlimited** (JSON cannot carry Infinity).
        teasers: { type: integer, description: "How many censored rows a gated caller gets (12). Absent for subscriber/admin." }
        explanations: { type: boolean }
        calculator: { type: boolean }
        applyLinks: { type: boolean }
        exportData: { type: boolean }
        admin: { type: boolean }
    Subscription:
      type: object
      required: [active, daysLeft]
      properties:
        status: { type: string, enum: [none, trial, active] }
        plan: { type: string, nullable: true }
        validUntil: { type: string, format: date-time, nullable: true }
        grantedBy: { type: string, nullable: true }
        grantedAt: { type: string, format: date-time, nullable: true }
        note: { type: string, nullable: true }
        active: { type: boolean, description: 'Computed by the server from `validUntil`.' }
        daysLeft: { type: integer, nullable: true, description: 'Whole days until `validUntil`; null when none.' }
    User:
      type: object
      description: The account summary. **Never** includes the password hash or salt.
      required: [id, username, role, createdAt, disabled, subscription, entitlements, hasProfile]
      properties:
        id: { type: string }
        username: { type: string }
        email: { type: string, nullable: true }
        company: { type: string, nullable: true }
        role: { type: string, enum: [admin, user] }
        createdAt: { type: string, format: date-time }
        lastLoginAt: { type: string, format: date-time, nullable: true }
        disabled: { type: boolean }
        subscription: { $ref: '#/components/schemas/Subscription' }
        entitlements: { $ref: '#/components/schemas/Entitlements' }
        hasProfile: { type: boolean }
...
    AuthActionResponse:
      type: object
      required: [success, user]
      properties:
        success: { type: boolean, enum: [true] }
        user: { $ref: '#/components/schemas/User' }
...